Move models to core module
