--- chip_logistics/models/__init__.py
@@ -1,1 +1,0 @@
-"""Module with application models."""
