+++ tests/articles/test_calcs.py
@@ -6,7 +6,7 @@ from decimal import Decimal
 import pytest
 
 from chip_logistics.core.articles.calcs import calculate_article_price
-from chip_logistics.models.articles import ArticleItem
+from chip_logistics.core.articles.models import ArticleItem
 from tests.articles.conftest import test_articles
